+++ src/Api/Discovery/DiscoveryManager.php
@@ -17,6 +17,8 @@ use Rhumsaa\Uuid\Uuid;
 use Webmozart\Expression\Expression;
 
 /**
+ * Manages the resource discovery of a Puli project.
+ *
  * @since  1.0
  * @author Bernhard Schussek <bschussek@gmail.com>
  */